Fixed Charachter Controller
I created a new exported var called snap so you can easily see this my snap length always was either 16 or 32 i tried both. but each block is a 16 by 16 block so when it was one block down it beleieved it was still snaped to the floor. the snapping was neccesary for the slopes. I changed the snap distance to 8 so the slopes still wokred because they are half of a block hence haklf of 16 is 8. This fixed the player movement